+++ tests/data/preview/return_annotation_brackets.py
@@ -91,6 +91,10 @@ def foo() -> tuple[int, int, int,]:
 def frobnicate() -> "ThisIsTrulyUnreasonablyExtremelyLongClassName | list[ThisIsTrulyUnreasonablyExtremelyLongClassName]":
     pass
 
+# Should not remove parens around walrus operator
+def walrus() -> (x := 1):
+    pass
+
 # output
 # Control
 def double(a: int) -> int:
@@ -219,4 +223,9 @@ def frobnicate() -> (
     "ThisIsTrulyUnreasonablyExtremelyLongClassName |"
     " list[ThisIsTrulyUnreasonablyExtremelyLongClassName]"
 ):
+    pass
+
+
+# Should not remove parens around walrus operator
+def walrus() -> (x := 1):
     pass